7. Go to https://rahulshettyacademy.com/locatorspractice/ & login
By inputting username and incorrect password, and try to get the error
message but as there is milisesond of delay in displaying the error
message, it will throw an error saying no such element: Unable to locate
element. To resolve this, we will use implicit wait mechanism in next
commit, the code for which will be placed just below invoking broser.