8. Implicit wait will trigger for max of given seconds when required.
Lecture 35: In this case, as 5 seconds is given, it will wait for 5
seconds if the driver  is not able to find the element for upto 5
seconds on each steps before moving forward. If the element doesn't show
till 5 seconds, it will fail the test. This is the use of implicit wait.