Lecture 39 - Section 6 - Removing element click Intercept error
So, this error will mostly occur in all the single page application.
Here in this case, while page shifted to another sub page, the website
link remain same, but the page changed to new form without changing the
webpage. Here the other form was already filled, when the form was
transiting from 1 page to another, which triggered the error by using
Thread.sleep(miliseconds) after clicking on forgot password. When the
application is in changing state, we might get this error until the view
get stable in single page application. Here we can also use explicit
wait, which will be discussed later. Remember, in implicit wait, we use
it if the element doesn't show, but here in this example, the elements
are already there, but they are in the changing state, that are not
properly clickable/view-able