Lecture 41 & 42 - Locators2
1. Get the login message using tagname - use this when there is unique
tagname present and gettext to validate the message and login using
testng
2. get and paste the testng dependency from mvn under dependencies in
pom.xml to update.
3. Remember to remove scope or comment out <scope> from testng to see
testng related files. if we keep <scope>, we will not see testng related
jars in actual program files or just comment it out by holding control +
command + /
4.Compare Acutual(that we got from element, (with) what we see expected,
as we saw on website)
5.Use, previous version of testng if the current one does not
automatically pull the jar and auto import the package.
6. Here, why we are inputting explicit wait or thread.sleep after log in
to get the message is there is already 1 p tag available after we click
on log in button, hence we want to explicitly wait for the p tag that
would appear upon successful log in. hence we're using explicit wait.
Hence implicit wait will not work here.
7. Make name dynamic by assigning a global variable at the top  named
String name
8. Identify element based upon its text (log out button)