Lecture 55,56
1. Click on from by id = ctl00_mainContent_ddl_originStation1_CTXT
2. Click on Banglore by xpath//a[@value='BLR']
3. After clicking on Banglore another drop down will open automatically
but there is delay of 1 or 2 seconds, hence thread.sleep
4. click on Chennai by xpath and using index (//a[@value='MAA'])[2],
this is because, first index will be from the departure and it will be
visibile in this case, hence, we'd have to select, 2nd index.
5. Another method to select BLR as using index is not a best opetion is
to get an xpath using parent child relatiionship.
6. In this case, get the xpath of the whole window of to/destination and
in that tab by putting a space write an xpath for MAA/Chennai without
index. the index of parent child relationship would be
//div[@id='glsctl00_mainContent_ddl_destinationStation1_CTNR']
//a[@value='MAA']
7.