Lecture 63,64, check if one way is selected by get attribute IMP
1. Check if 1 way is selected id - ctl00_mainContent_rbtnl_Trip_0
2. Click on round trip, id - ctl00_mainContent_rbtnl_Trip_1
3. Select current date by css .class, remember to remove any space if
class name is combination of 2 class, class of current date will be
unique .ui-state-default.ui-state-highlight
4. Check if the return day section is enabled or disabled, but isEnable
will not work as if we click on return day, it will be highlighted and
the radio button will automatically shift to round trip as the website
is designed dynamically.
5. To resolve this, in round trip, check the html code and take a
screenshot, now select one way and and take another screenshot of the
html code for that div/section, one attribute will be changed. Here in
this case, it is opacity which changed from 1 to 0.5 in style attribute