L-101 - Scroll  to get view of element to perform action
url - https://rahulshettyacademy.com/AutomationPractice/
1. While performing the tests in headless mode which nowadays every
company is doing it, sometimes the test will fail as the webelement is
not in visible mode, for which we'd have to scroll the web page to have
that in the visible mode before performing any action.
2. count the total of web table fixed header for which we'd have to
apply scroll
3. By default selenium will not provide any scrolling techniques, so,
we'd have to use javascript.
4.To execute javascript, we'd have to create javascript executor object,
which will help you to execute java script
5. cast driver to javacriptExecutor
6. to execute get script >> console window.scroll(0,500)
7. js.executescript(window.scroll(0,500) - to scroll page
8. js.executescript("document.querryselector('locator').scrollTop=5000")